axf/requirement.txt
django==2.0.7
pymysql==0.9.3
djangorestframework==3.9.3
django-filter==2.1.0
django-redis==4.10.0
redis==3.2.1
django-cors-headers==2.5.3